requirements.txt
astrapy
langchain_openai
ftfy
bs4
python-dotenv